Modal for View Recipe option on cards
Made a modal appear when clicking the "Ver receta" button in the recipe cards in index.html, modal is in the same file.

Careful with the id="viewRecipeModal" in the outter tag of the modal, this links to the <a> tag in recipe_card.html.

Also added some styles in style.css for the headings in the modal.