View recipe layout change
Changed the image size in view recipy for the modal and adjusted the position of the button "cerrar".

Added the style for the image in the modal in style.css. Also changed the btn-primary focus color so it won't be bootstrap's blue.

Also changed the spinner's color in the index_admin,html file